Add TEMP_IS_FAHRENHEIT
Can be set to false to choose celsius. Could not locate a source for a heatindex formula for C, so we convert to F and then back to C inside the heatindex function. We do this using preprocessor defines, so it has no performance impact on F users. Sorry, C users.